Reorganize command options in Breeze
The command options in breeze have been rather badly organized. The
"common_option" in utils package was misplaced and it contained
a lot of options that were not really "common" - only used in
one or two commands.

This PR reorganizes those options by making a bit more reasonable
decisions where options are placed:

* when options are used by a single command - they are defined in that
  command
* when options are used by more comnands in the same command module,
  they are defined in that command module - at the tops
* in cases of "image" and "package_installation" optins - they are
  used by several modules only and they have their own "common_"*"
  modules
* only in case a command is used in multiple modules, they are
  put in "common_options".
* the "common_options" module has been moved to "commands" package
  (it makes much more sense to be there - together with all commands
* the command options are sorted alphabetically in their corresponding
  files.
* Fixed missing default for "--platform" flag
