+++ dev/breeze/src/airflow_breeze/commands/kubernetes_commands.py
@@ -28,12 +28,7 @@ from shlex import quote
 
 import click
 
-from airflow_breeze.commands.production_image_commands import run_build_production_image
-from airflow_breeze.global_constants import ALLOWED_EXECUTORS, ALLOWED_KUBERNETES_VERSIONS
-from airflow_breeze.params.build_prod_params import BuildProdParams
-from airflow_breeze.utils.ci_group import ci_group
-from airflow_breeze.utils.click_utils import BreezeGroup
-from airflow_breeze.utils.common_options import (
+from airflow_breeze.commands.common_options import (
     option_debug_resources,
     option_dry_run,
     option_include_success_outputs,
@@ -44,6 +39,11 @@ from airflow_breeze.utils.common_options import (
     option_skip_cleanup,
     option_verbose,
 )
+from airflow_breeze.commands.production_image_commands import run_build_production_image
+from airflow_breeze.global_constants import ALLOWED_EXECUTORS, ALLOWED_KUBERNETES_VERSIONS
+from airflow_breeze.params.build_prod_params import BuildProdParams
+from airflow_breeze.utils.ci_group import ci_group
+from airflow_breeze.utils.click_utils import BreezeGroup
 from airflow_breeze.utils.console import Output, get_console
 from airflow_breeze.utils.custom_param_types import CacheableChoice, CacheableDefault
 from airflow_breeze.utils.kubernetes_utils import (
@@ -111,16 +111,18 @@ option_executor = click.option(
     default=CacheableDefault(ALLOWED_EXECUTORS[0]),
     envvar="EXECUTOR",
 )
-
-option_kubernetes_version = click.option(
-    "--kubernetes-version",
-    help="Kubernetes version used to create the KinD cluster of.",
-    type=CacheableChoice(ALLOWED_KUBERNETES_VERSIONS),
-    show_default=True,
-    default=CacheableDefault(ALLOWED_KUBERNETES_VERSIONS[0]),
-    envvar="KUBERNETES_VERSION",
+option_force_recreate_cluster = click.option(
+    "--force-recreate-cluster",
+    help="Force recreation of the cluster even if it is already created.",
+    is_flag=True,
+    envvar="FORCE_RECREATE_CLUSTER",
 )
-
+option_force_venv_setup = click.option(
+    "--force-venv-setup",
+    help="Force recreation of the virtualenv.",
+    is_flag=True,
+    envvar="FORCE_VENV_SETUP",
+)
 option_image_tag = click.option(
     "-t",
     "--image-tag",
@@ -129,58 +131,14 @@ option_image_tag = click.option(
     show_default=True,
     envvar="IMAGE_TAG",
 )
-
-option_wait_time_in_seconds = click.option(
-    "--wait-time-in-seconds",
-    help="Wait for Airflow webserver for specified number of seconds.",
-    type=click.IntRange(0),
-    default=120,
-    envvar="WAIT_TIME_IN_SECONDS",
+option_kubernetes_version = click.option(
+    "--kubernetes-version",
+    help="Kubernetes version used to create the KinD cluster of.",
+    type=CacheableChoice(ALLOWED_KUBERNETES_VERSIONS),
+    show_default=True,
+    default=CacheableDefault(ALLOWED_KUBERNETES_VERSIONS[0]),
+    envvar="KUBERNETES_VERSION",
 )
-
-option_wait_time_in_seconds_0_default = click.option(
-    "--wait-time-in-seconds",
-    help="Wait for Airflow webserver for specified number of seconds.",
-    type=click.IntRange(0),
-    default=0,
-    envvar="WAIT_TIME_IN_SECONDS",
-)
-
-option_force_recreate_cluster = click.option(
-    "--force-recreate-cluster",
-    help="Force recreation of the cluster even if it is already created.",
-    is_flag=True,
-    envvar="FORCE_RECREATE_CLUSTER",
-)
-
-option_force_venv_setup = click.option(
-    "--force-venv-setup",
-    help="Force recreation of the virtualenv.",
-    is_flag=True,
-    envvar="FORCE_VENV_SETUP",
-)
-
-option_use_standard_naming = click.option(
-    "--use-standard-naming",
-    help="Use standard naming.",
-    is_flag=True,
-    envvar="USE_STANDARD_NAMING",
-)
-
-option_multi_namespace_mode = click.option(
-    "--multi-namespace-mode",
-    help="Use multi namespace mode.",
-    is_flag=True,
-    envvar="MULTI_NAMESPACE_MODE",
-)
-
-option_rebuild_base_image = click.option(
-    "--rebuild-base-image",
-    help="Rebuilds base Airflow image before building K8S image.",
-    is_flag=True,
-    envvar="REBUILD_BASE_IMAGE",
-)
-
 option_kubernetes_versions = click.option(
     "--kubernetes-versions",
     help="Kubernetes versions used to run in parallel (space separated).",
@@ -189,14 +147,44 @@ option_kubernetes_versions = click.option(
     default=" ".join(ALLOWED_KUBERNETES_VERSIONS),
     envvar="KUBERNETES_VERSIONS",
 )
-
+option_multi_namespace_mode = click.option(
+    "--multi-namespace-mode",
+    help="Use multi namespace mode.",
+    is_flag=True,
+    envvar="MULTI_NAMESPACE_MODE",
+)
+option_rebuild_base_image = click.option(
+    "--rebuild-base-image",
+    help="Rebuilds base Airflow image before building K8S image.",
+    is_flag=True,
+    envvar="REBUILD_BASE_IMAGE",
+)
 option_upgrade = click.option(
     "--upgrade",
     help="Upgrade Helm Chart rather than installing it.",
     is_flag=True,
     envvar="UPGRADE",
 )
-
+option_use_standard_naming = click.option(
+    "--use-standard-naming",
+    help="Use standard naming.",
+    is_flag=True,
+    envvar="USE_STANDARD_NAMING",
+)
+option_wait_time_in_seconds = click.option(
+    "--wait-time-in-seconds",
+    help="Wait for Airflow webserver for specified number of seconds.",
+    type=click.IntRange(0),
+    default=120,
+    envvar="WAIT_TIME_IN_SECONDS",
+)
+option_wait_time_in_seconds_0_default = click.option(
+    "--wait-time-in-seconds",
+    help="Wait for Airflow webserver for specified number of seconds.",
+    type=click.IntRange(0),
+    default=0,
+    envvar="WAIT_TIME_IN_SECONDS",
+)
 option_parallelism_cluster = click.option(
     "--parallelism",
     help="Maximum number of processes to use while running the operation in parallel for cluster operations.",
